fix/map-property-card-padding
I removed the padding around each of the individual property cards and added padding to the left and right of the parent container that is equal to the margin for each property card. Doing so makes the spaces around and between each property card the same and still takes up the same total space on the page.