chore: made credentials optional for current/old pipeline
The pipeline can be run locally but will not write changes to the bucket
and will not report errors to slack without credentials. Previously, the
code would error when credentials were missing.

Added BACKUP_SCHEMA=False and made FORCE_RELOAD=True as the default.
These are temporary fixes and will not be relevant once postgres is
removed.